+++ src/utils/hash.js
