+++ src/utils/sccache.js
@@ -2,7 +2,7 @@ const childProcess = require('child_process');
 const os = require('os');
 const path = require('path');
 
-const { color } = require('./e-utils')
+const { color } = require('./logging')
 
 const getExternalBinaries = root => path.resolve(root, 'src', 'electron', 'external_binaries');
 
@@ -37,8 +37,6 @@ function ensureSCCache(config) {
 }
 
 module.exports = {
-  sccache: {
-    ensure: ensureSCCache,
-    exec: root => getSCCacheExec(root),
-  }
+  ensure: ensureSCCache,
+  exec: root => getSCCacheExec(root),
 }
